requirements.txt
Django==2.1.5
ujson==1.35
python-slugify==2.0.1
social-auth-app-django==3.1.0

ipython
